feat: minify option for resourcePot and module
