+++ crates/compiler/tests/fixtures/minify/css/basic/output.js
@@ -2,9 +2,5 @@
  (globalThis || window || global)['__farm_default_namespace__'] = {__FARM_TARGET_ENV__: 'browser'};(function(r,e){var t={};function n(r){return Promise.resolve(o(r))}function o(e){if(t[e])return t[e].exports;var i={id:e,exports:{}};r[e](i,i.exports,o,n);t[e]=i;return i.exports}o(e)})({"ec853507":function  (_,e,l,n){console.log("runtime/index.js")(globalThis||window||global).__farm_default_namespace__.__farm_module_system__.setPlugins([]);},},"ec853507");(function(_){for(var r in _){_[r].__farm_resource_pot__='index_dcdc.js';(globalThis || window || global)['__farm_default_namespace__'].__farm_module_system__.register(r,_[r])}})({"b5d64806":function  (e,t,u,c){"use strict";Object.defineProperty(t,"__esModule",{value:!0});},});(globalThis || window || global)['__farm_default_namespace__'].__farm_module_system__.setInitialLoadedResources([]);(globalThis || window || global)['__farm_default_namespace__'].__farm_module_system__.setDynamicModuleResourcesMap({  });var farmModuleSystem = (globalThis || window || global)['__farm_default_namespace__'].__farm_module_system__;farmModuleSystem.bootstrap();var entry = farmModuleSystem.require("b5d64806");
 
 //index_6dd2.css:
- #root {
-  color: green;
-}
-.body {
-  color: red;
-}
+ #root{color:green}
+.body{color:red}
