+++ crates/plugin_css/Cargo.toml
@@ -15,7 +15,6 @@ farmfe_toolkit = { path = "../toolkit", version = "0.0.2" }
 farmfe_utils = { path = "../utils", version = "0.1.3" }
 farmfe_macro_cache_item = { path = "../macro_cache_item", version = "0.1.2" }
 rkyv = { version = "0.7.42" }
-farmfe_plugin_minify = { path = "../plugin_minify" }
 
 [dev-dependencies]
 farmfe_testing_helpers = { path = "../testing_helpers", version = "0.0.2" }
