+++ crates/plugin_html/Cargo.toml
@@ -13,4 +13,3 @@ documentation = "https://docs.rs/farmfe_plugin_html"
 farmfe_core = { path = "../core", version = "0.4.1" }
 farmfe_toolkit = { path = "../toolkit", version = "0.0.2" }
 farmfe_testing_helpers = { path = "../testing_helpers", version = "0.0.2" }
-farmfe_plugin_minify = { path = "../plugin_minify/" }
