+++ crates/plugin_runtime/Cargo.toml
@@ -14,4 +14,3 @@ farmfe_core = { path = "../core", version = "0.4.1" }
 farmfe_toolkit = { path = "../toolkit", version = "0.0.2" }
 farmfe_testing_helpers = { path = "../testing_helpers", version = "0.0.2" }
 farmfe_utils = { path = "../utils", version = "0.1.3" }
-farmfe_plugin_minify = { path = "../plugin_minify", version = "0.0.1" }
